+++ backend/services/auth-service-backend/src/main/java/br/com/condominium/auth/model/repository/CredentialRepository.java
@@ -1,6 +1,6 @@
-package br.com.condominium.auth.credential.repository;
+package br.com.condominium.auth.model.repository;
 
-import br.com.condominium.auth.credential.model.Credential;
+import br.com.condominium.auth.model.entity.Credential;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;